add java_notes

+++ day7/TestJieCheng.java
@@ -2,8 +2,8 @@ package day7;
 
 public class TestJieCheng {
 	public static void main(String[] args) {
-		int sum = 1;			//每个数字自个的阶乘
-		int total =0;			//阶乘之和
+		long sum = 1;			//每个数字自个的阶乘
+		long total =0;			//阶乘之和
 		for(int i=1;i<=20;i++) {
 			for(int j=1;j<=i;j++) {
 				sum *= j;
